feat: API Rest Product works and validate

+++ backend/src/main/java/com/hexagonal/ecommerce/infrastructure/database/mysql/adapter/ProductRepositoryAdapter.java
@@ -1,18 +1,21 @@
 package com.hexagonal.ecommerce.infrastructure.database.mysql.adapter;
 
+import java.util.NoSuchElementException;
+
+import org.springframework.stereotype.Component;
+import org.springframework.transaction.annotation.Transactional;
 
 import com.hexagonal.ecommerce.domain.model.Product;
 import com.hexagonal.ecommerce.domain.repository.ProductRepositoryPort;
 import com.hexagonal.ecommerce.infrastructure.database.mysql.entity.ProductHexaEntity;
 import com.hexagonal.ecommerce.infrastructure.database.mysql.mapper.ProductEntityMapper;
 import com.hexagonal.ecommerce.infrastructure.database.mysql.repository.ProductJpaRepository;
+
 import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Component;
 
-import java.util.NoSuchElementException;
-
-@RequiredArgsConstructor
 @Component
+@Transactional
+@RequiredArgsConstructor
 public class ProductRepositoryAdapter implements ProductRepositoryPort {
 
     private final ProductJpaRepository productJpaRepository;
@@ -34,17 +37,27 @@ public class ProductRepositoryAdapter implements ProductRepositoryPort {
 
     @Override
     public void deleteProductByProductId(String productId) {
+        if (productJpaRepository.findByProductId(productId).isEmpty()) {
+            throw new IllegalArgumentException("Product not exists");
+        }
         productJpaRepository.deleteByProductId(productId);
     }
 
     @Override
     public Product createProduct(ProductHexaEntity productHexaEntity) {
+        if (!productJpaRepository.findByProductId(productHexaEntity.getProductId()).isEmpty()) {
+            throw new IllegalArgumentException("Product already exists");
+        }
         ProductHexaEntity saved = productJpaRepository.save(productHexaEntity);
         return ProductEntityMapper.fromEntityToProduct(saved);
     }
 
     @Override
     public Product updateProduct(ProductHexaEntity productHexaEntity) {
+        if (productJpaRepository.findByProductId(productHexaEntity.getProductId()).isEmpty()) {
+            throw new IllegalArgumentException("Product not exists");
+        }
+        productJpaRepository.deleteByProductId(productHexaEntity.getProductId());
         ProductHexaEntity updated = productJpaRepository.save(productHexaEntity);
         return ProductEntityMapper.fromEntityToProduct(updated);
     }
